Move more code to the core package
