+++ packages/core/src/helpers/github/events.ts
@@ -5,18 +5,22 @@ import {
   EnhancedGitHubEvent,
   GitHubEventAction,
   GitHubEventSubjectType,
+  GitHubIcon,
+  GitHubIssue,
   GitHubPullRequest,
   MultipleStarEvent,
+  ThemeColors,
 } from '../../types'
-import { getBranchNameFromRef, isDraft, isPullRequest } from './shared'
-
-export function getOlderEventDate(
-  events: EnhancedGitHubEvent[],
-  field: keyof EnhancedGitHubEvent = 'created_at',
-) {
-  const olderItem = sortEvents(events, field, 'desc').pop()
-  return olderItem && olderItem[field]
-}
+import {
+  getBranchNameFromRef,
+  getCommitIconAndColor,
+  getIssueIconAndColor,
+  getPullRequestIconAndColor,
+  getReleaseIconAndColor,
+  getTagIconAndColor,
+  isDraft,
+  isPullRequest,
+} from './shared'
 
 export const eventActions: GitHubEventAction[] = [
   'added',
@@ -34,6 +38,27 @@ export const eventActions: GitHubEventAction[] = [
   'starred',
   'updated',
 ]
+
+export const eventSubjectTypes: GitHubEventSubjectType[] = [
+  'Branch',
+  'Commit',
+  'Issue',
+  'PullRequest',
+  'PullRequestReview',
+  'Release',
+  'Repository',
+  'Tag',
+  'User',
+  'Wiki',
+]
+
+export function getOlderEventDate(
+  events: EnhancedGitHubEvent[],
+  field: keyof EnhancedGitHubEvent = 'created_at',
+) {
+  const olderItem = sortEvents(events, field, 'desc').pop()
+  return olderItem && olderItem[field]
+}
 
 export function getEventActionMetadata<T extends GitHubEventAction>(
   action: T,
@@ -634,4 +659,217 @@ export function sortEvents(
     .uniqBy('id')
     .orderBy(field, order)
     .value()
+}
+
+export function getEventSubjectType(
+  event: EnhancedGitHubEvent,
+): GitHubEventSubjectType | null {
+  if (!(event && event.type)) return null
+
+  switch (event.type) {
+    case 'CommitCommentEvent':
+      return 'Commit'
+
+    case 'CreateEvent':
+    case 'DeleteEvent': {
+      switch (event.payload.ref_type) {
+        case 'repository':
+          return 'Repository'
+        case 'branch':
+          return 'Branch'
+        case 'tag':
+          return 'Tag'
+        default:
+          return null
+      }
+    }
+
+    case 'ForkEvent':
+      return 'Repository'
+    case 'GollumEvent':
+      return 'Wiki'
+    case 'IssueCommentEvent':
+      return 'Issue'
+    case 'IssuesEvent':
+      return 'Issue'
+    case 'MemberEvent':
+      return 'User'
+    case 'PublicEvent':
+      return 'Repository'
+    case 'PullRequestEvent':
+      return 'PullRequest'
+    case 'PullRequestReviewCommentEvent':
+      return 'PullRequestReview'
+    case 'PullRequestReviewEvent':
+      return 'PullRequestReview'
+    case 'PushEvent':
+      return 'Commit'
+    case 'ReleaseEvent':
+      return 'Release'
+    case 'WatchEvent':
+    case 'WatchEvent:OneUserMultipleRepos':
+      return 'Repository'
+
+    default:
+      return null
+  }
+}
+
+export function getEventIconAndColor(
+  event: EnhancedGitHubEvent,
+): { color?: keyof ThemeColors; icon: GitHubIcon; subIcon?: GitHubIcon } {
+  switch (event.type) {
+    case 'CommitCommentEvent':
+      return {
+        ...getCommitIconAndColor(),
+        subIcon: 'comment-discussion',
+      }
+
+    case 'CreateEvent': {
+      switch (event.payload.ref_type) {
+        case 'repository':
+          return { icon: 'repo' }
+        case 'branch':
+          return { icon: 'git-branch' }
+        case 'tag':
+          return { icon: 'tag' }
+        default:
+          return { icon: 'plus' }
+      }
+    }
+
+    case 'DeleteEvent': {
+      switch (event.payload.ref_type) {
+        case 'repository':
+          return { icon: 'repo', color: 'red' }
+        case 'branch':
+          return { icon: 'git-branch', color: 'red' }
+        case 'tag':
+          return { icon: 'tag', color: 'red' }
+        default:
+          return { icon: 'trashcan' }
+      }
+    }
+
+    case 'ForkEvent':
+      return { icon: 'repo-forked' }
+
+    case 'GollumEvent':
+      return { icon: 'book' }
+
+    case 'IssueCommentEvent': {
+      return {
+        ...(isPullRequest(event.payload.issue)
+          ? getPullRequestIconAndColor(event.payload.issue as GitHubPullRequest)
+          : getIssueIconAndColor(event.payload.issue)),
+        subIcon: 'comment-discussion',
+      }
+    }
+
+    case 'IssuesEvent': {
+      const issue = event.payload.issue
+
+      switch (event.payload.action) {
+        case 'opened':
+          return getIssueIconAndColor({ state: 'open' } as GitHubIssue)
+        case 'closed':
+          return getIssueIconAndColor({ state: 'closed' } as GitHubIssue)
+
+        case 'reopened':
+          return {
+            ...getIssueIconAndColor({ state: 'open' } as GitHubIssue),
+            icon: 'issue-reopened',
+          }
+        // case 'assigned':
+        // case 'unassigned':
+        // case 'labeled':
+        // case 'unlabeled':
+        // case 'edited':
+        // case 'milestoned':
+        // case 'demilestoned':
+        default:
+          return getIssueIconAndColor(issue)
+      }
+    }
+    case 'MemberEvent':
+      return { icon: 'person' }
+
+    case 'PublicEvent':
+      return { icon: 'globe', color: 'blue' }
+
+    case 'PullRequestEvent': {
+      const pullRequest = event.payload.pull_request
+
+      switch (event.payload.action) {
+        case 'opened':
+        case 'reopened':
+          return getPullRequestIconAndColor({
+            draft: pullRequest.draft,
+            state: 'open',
+            merged: false,
+            merged_at: undefined,
+            mergeable_state: pullRequest.mergeable_state,
+          })
+        // case 'closed': return getPullRequestIconAndColor({ state: 'closed' } as GitHubPullRequest);
+
+        // case 'assigned':
+        // case 'unassigned':
+        // case 'labeled':
+        // case 'unlabeled':
+        // case 'edited':
+        default:
+          return getPullRequestIconAndColor(pullRequest)
+      }
+    }
+
+    case 'PullRequestReviewCommentEvent':
+    case 'PullRequestReviewEvent': {
+      return {
+        ...getPullRequestIconAndColor(event.payload.pull_request),
+        subIcon: 'comment-discussion',
+      }
+    }
+
+    case 'PushEvent':
+      return { icon: 'code' }
+
+    case 'ReleaseEvent':
+      return isTagMainEvent(event)
+        ? getTagIconAndColor()
+        : getReleaseIconAndColor()
+
+    case 'WatchEvent':
+    case 'WatchEvent:OneUserMultipleRepos':
+      return { icon: 'star', color: 'yellow' }
+
+    default: {
+      const message = `Unknown event type: ${(event as any).type}`
+      console.error(message)
+      return { icon: 'mark-github' }
+    }
+  }
+}
+
+export function mergeEventsPreservingEnhancement(
+  newItems: EnhancedGitHubEvent[],
+  prevItems: EnhancedGitHubEvent[],
+) {
+  return sortEvents(
+    _.uniqBy(_.concat(newItems, prevItems), 'id').map(item => {
+      const newItem = newItems.find(i => i.id === item.id)
+      const existingItem = prevItems.find(i => i.id === item.id)
+      if (!(newItem && existingItem)) return item
+
+      const mergedItem = {
+        forceUnreadLocally: existingItem.forceUnreadLocally,
+        last_read_at: existingItem.last_read_at,
+        last_unread_at: existingItem.last_unread_at,
+        saved: existingItem.saved,
+        unread: existingItem.unread,
+        ...newItem,
+      }
+
+      return _.isEqual(mergedItem, existingItem) ? existingItem : mergedItem
+    }),
+  )
 }
