+++ packages/core/src/helpers/github/issues.ts
@@ -4,13 +4,57 @@ import _ from 'lodash'
 import {
   EnhancedGitHubIssueOrPullRequest,
   EnhancementCache,
+  GitHubIssue,
   GitHubIssueOrPullRequest,
   GitHubIssueOrPullRequestSubjectType,
+  GitHubPullRequest,
   IssueOrPullRequestColumnSubscription,
   IssueOrPullRequestPayloadEnhancement,
 } from '../../types'
-import { getOwnerAndRepo } from './shared'
+import {
+  getIssueIconAndColor,
+  getOwnerAndRepo,
+  getPullRequestIconAndColor,
+} from './shared'
 import { getRepoFullNameFromUrl } from './url'
+
+export const issueOrPullRequestSubjectTypes: GitHubIssueOrPullRequestSubjectType[] = [
+  'Issue',
+  'PullRequest',
+]
+
+export function getIssueOrPullRequestIconAndColor(
+  type: GitHubIssueOrPullRequestSubjectType,
+  issueOrPullRequest: GitHubIssueOrPullRequest,
+) {
+  return type === 'PullRequest'
+    ? getPullRequestIconAndColor(issueOrPullRequest as GitHubPullRequest)
+    : getIssueIconAndColor(issueOrPullRequest as GitHubIssue)
+}
+
+export function mergeIssuesOrPullRequestsPreservingEnhancement(
+  newItems: EnhancedGitHubIssueOrPullRequest[],
+  prevItems: EnhancedGitHubIssueOrPullRequest[],
+) {
+  return sortIssuesOrPullRequests(
+    _.uniqBy(_.concat(newItems || [], prevItems || []), 'id').map(item => {
+      const newItem = (newItems || []).find(i => i.id === item.id)
+      const existingItem = prevItems.find(i => i.id === item.id)
+      if (!(newItem && existingItem)) return item
+
+      const mergedItem = {
+        forceUnreadLocally: existingItem.forceUnreadLocally,
+        last_read_at: existingItem.last_read_at,
+        last_unread_at: existingItem.last_unread_at,
+        saved: existingItem.saved,
+        unread: existingItem.unread,
+        ...newItem,
+      }
+
+      return _.isEqual(mergedItem, existingItem) ? existingItem : mergedItem
+    }),
+  )
+}
 
 export function getIssueOrPullRequestSubjectType(
   item: GitHubIssueOrPullRequest,
@@ -185,7 +229,7 @@ export function getGitHubIssueSearchQuery(
   if (repoFullName) queryArr.push(`repo:${repoFullName}`)
 
   if (subjectType === 'Issue') queryArr.push('is:issue')
-  if (subjectType === 'PullRequest') queryArr.push('is:pr')
+  else if (subjectType === 'PullRequest') queryArr.push('is:pr')
 
   return queryArr.join(' ')
 }
