+++ packages/core/src/helpers/index.ts
@@ -1,3 +1,4 @@
+export * from './filters'
 export * from './github'
 export * from './graphql'
 export * from './shared'
